+++ packages/electron/electron-env.d.ts
@@ -2,7 +2,7 @@
 declare namespace NodeJS {
   interface ProcessEnv {
     NODE_ENV: 'development' | 'production'
-    readonly VITE_DEV_SERVER_HOST: string
+    readonly VITE_DEV_SERVER_HOSTNAME: string
     readonly VITE_DEV_SERVER_PORT: string
     readonly VITE_DEV_SERVER_URL: string
   }